Created Testcases Extracted KeyStorage Base to core Fixed some warnings
